third_party/WebKit/LayoutTests/editing/selection/leak-document-with-selection-inside.html

<html>
<body>
    <script id='targetHTML' type='text/html'>
        <span id='t'></span>
    </script>
    <script id='targetJS' type='text/html'>
        var element = document.getElementById("t");
        window.getSelection().setPosition(element, 0);
    </script>

    <script src="../../fast/js/resources/js-test-pre.js"></script>
    <script src="../../fast/dom/resources/leak-check.js"></script>
    <script>
        description("This test asserts that document doesn't leak when a selection is created inside the document.");

        var target = grabScriptText('targetHTML') + 
                     '<script>'+grabScriptText('targetJS')+'<'+'/script>';
        var tolerance = {'numberOfLiveDocuments': 0};

        doLeakTest(htmlToUrl(target), tolerance);
    </script>
    <script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
